feat: T-24 AI面试会话后端 - AIInterviewSession实体/Controller/Service/DTOs
- AIInterviewSession实体：会话状态流转(Pending/InProgress/Completed/Cancelled/NoShow)
- AIInterviewSessionsController: 7个端点(列表/详情/创建/开始/完成/取消/提交回答/下一题)
- AIInterviewSessionService: Mock问题库(5题)+AI回答模拟
- InterviewInvitation实体: Id改为BIGINT(匹配EF Core)
- MySQL建表: InterviewInvitations + AIInterviewSessions(含外键)
- SeedData: AIInterviewSession种子数据

+++ backend/TalentPilot.Api/Models/Entities/InterviewInvitation.cs
@@ -11,13 +11,13 @@ public class InterviewInvitation
     public int Id { get; set; }
 
     [Column("CandidateId")]
-    public int CandidateId { get; set; }
+    public long CandidateId { get; set; }
 
     [Column("JobPostId")]
     public int JobPostId { get; set; }
 
     [Column("InvitedByUserId")]
-    public int InvitedByUserId { get; set; }
+    public long InvitedByUserId { get; set; }
 
     [Column("InterviewTime")]
     public DateTime? InterviewTime { get; set; }
